+++ client/src/components/Overview/ProductInfo.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import local from '../../styles/Overview.css';
+import local from '../../styles/Overview/ProductInfo.css';
 
 function ProductInfo({
   current, price, origPrice, onSale,
