+++ client/src/components/Overview/StyleEntry.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import local from '../../styles/Overview.css';
+import local from '../../styles/Overview/StyleEntry.css';
 import { buildHandleEnterKeyPress } from '../../util';
 
 function StyleEntry({
