--- client/src/styles/Overview.css
@@ -1,73 +1,0 @@
-.overview {
-  background: #BFF;
-}
-
-.overviewMain {
-  display: flex;
-  width: 100%;
-  margin: auto;
-}
-
-.infoStylesCart {
-  width: 35%;
-}
-
-.gallery {
-  width: 65%;
-}
-
-.galleryMain {
-  display: block;
-}
-
-.galleryThumbs {
-  display: block;
-}
-
-.productInfo {
-  display: block;
-}
-
-.productDesc {
-  display: block;
-}
-
-.cart {
-  display: block;
-}
-
-.styles {
-  display: flex;
-  max-width: 281px;
-  flex-wrap: wrap;
-}
-
-.style {
-  display: inline-block;
-  border: 1px solid #111;
-  border-radius: 50%;
-  min-width: max(22%, 35px);
-  max-width: 65px;
-  aspect-ratio: 1 / 1;
-  margin: 1.5px;
-  background-size: cover;
-  text-align: right;
-  font-size: 1.3rem;
-  font-weight: bold;
-  color: #111;
-  text-shadow: 0px 0px 2.5px white;
-  vertical-align: text-top;
-  -webkit-text-stroke: 1px whitesmoke;
-}
-
-.styleHead {
-  text-transform: uppercase;
-}
-
-.styleTitle {
-  font-weight: bold;
-}
-
-.oldPrice {
-  text-decoration: line-through;
-}
