+++ client/src/styles/Overview/Cart.css
@@ -1,0 +1,3 @@
+.cart {
+  display: block;
+}
