+++ client/src/styles/Overview/Gallery.css
@@ -1,0 +1,11 @@
+.gallery {
+  width: 65%;
+}
+
+.galleryMain {
+  display: block;
+}
+
+.galleryThumbs {
+  display: block;
+}
