+++ client/src/styles/Overview/ProductDesc.css
@@ -1,0 +1,3 @@
+.productDesc {
+  display: block;
+}
