+++ client/src/styles/Overview/ProductInfo.css
@@ -1,0 +1,7 @@
+.productInfo {
+  display: block;
+}
+
+.oldPrice {
+  text-decoration: line-through;
+}
