+++ client/src/styles/Overview/StyleEntry.css
@@ -1,0 +1,17 @@
+.style {
+  display: inline-block;
+  border: 1px solid #111;
+  border-radius: 50%;
+  min-width: max(22%, 35px);
+  max-width: 65px;
+  aspect-ratio: 1 / 1;
+  margin: 1.5px;
+  background-size: cover;
+  text-align: right;
+  font-size: 1.3rem;
+  font-weight: bold;
+  color: #111;
+  text-shadow: 0px 0px 2.5px white;
+  vertical-align: text-top;
+  -webkit-text-stroke: 1px whitesmoke;
+}
